FEAT: create class Game
